\name{CMUDict}
\docType{data}
\alias{CMUDict}
\title{CMU Pronouncing Dictionary}
\description{
	Gives all the entries in the CMU Pronouncing dictionary.
}
\format{
	An environment containing all the entries in the CMU Pronouncing dictionary. The environment is used as a hashtable.
}
\source{CMU Pronouncing Dictionary}
\references{ http://www.speech.cs.cmu.edu/cgi-bin/cmudict}
\keyword{datasets}